+++ context/app/test_routes_main.py
@@ -65,28 +65,6 @@ def mock_get(path, **kwargs):
     return MockResponse()
 
 
-class MockSearch():
-    def __init__(self, **kwargs):
-        pass
-
-    def query(self, search, **kwargs):
-        pass
-
-    def execute(self):
-        return [MockResponse()]
-
-
-class MockResponse():
-    def __init__(self):
-        pass
-
-    def to_dict(self):
-        return {
-            'provenance_create_timestamp': '100000',
-            'provenance_modified_timestamp': '100000',
-        }
-
-
 @pytest.mark.parametrize(
     'path',
     ['/', '/help']
@@ -95,7 +73,6 @@ class MockResponse():
 )
 def test_200_html_page(client, path, mocker):
     mocker.patch('requests.get', side_effect=mock_get)
-    mocker.patch.object(api_client, 'Search', MockSearch)
     response = client.get(path)
     assert response.status == '200 OK'
     assert_is_valid_html(response)
@@ -105,8 +82,7 @@ def test_200_html_page(client, path, mocker):
     'path',
     [f'/browse/{t}/fake-uuid.json' for t in types]
 )
-def test_200_json_page(client, path, mocker):
-    mocker.patch.object(api_client, 'Search', MockSearch)
+def test_200_json_page(client, path):
     response = client.get(path)
     assert response.status == '200 OK'
     json.loads(response.data.decode('utf8'))