+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Hit Elasticsearch instance and start configuring facets.
 - Add react, webpack, eslint, and styling dependencies.
+- Make the Elasticsearch endpoint part of the config.
 
 ## [v0.0.10](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.10) - 2020/03/06
 ### Added